Layout footer, header 생성
