--- src/main/java/com/lionheart15/ideaboard/domain/entity/Like.java
@@ -1,24 +1,0 @@
-package com.lionheart15.ideaboard.domain.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-public class Like {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private Long user_id;
-    private Long content_id;
-}
