+++ src/main/resources/templates/layouts/footer.mustache
@@ -1,1 +1,6 @@
-푸터
+<div class="container-fluid">
+    <hr>
+    <p>Copyright lionheart-15 <a href="/privacy">Privacy</a> <a href="/term">Terms</a></p>
+</div>
+</body>
+</html>
